+++ javascript/src/main/kotlin/Rect.kt
@@ -1,6 +1,6 @@
 data class Rect(
-    val left: Double,
-    val top: Double,
-    val right: Double,
-    val bottom: Double
+    val left: Float,
+    val top: Float,
+    val right: Float,
+    val bottom: Float
 )
